Specify logo dimensions
I've fixed the heights at 160 for gold, 150 for silver to make the gold stand out a little.
I've fixed the widths based on the ratio for the height.
I've fixed the dimensions in general to avoid having one massive logo and a bunch of tiny ones.
Yes, I should've probably actually resized the images. No, I'm not doing that right now, as they're not too far off.